requirements.txt
streamlit
networkx
matplotlib
pyvis
termcolor
PyPDF2
xmltodict
openai
python-dotenv